+++ packages/core/src/components/Form.tsx
@@ -9,21 +9,17 @@ import {
   FieldPathId,
   FieldPathList,
   FormContextType,
-  GenericObjectType,
   getChangedFields,
   getTemplate,
   getUiOptions,
   isObject,
   mergeObjects,
-  NAME_KEY,
-  PathSchema,
   StrictRJSFSchema,
   Registry,
   RegistryFieldsType,
   RegistryWidgetsType,
   RJSFSchema,
   RJSFValidationError,
-  RJSF_ADDITIONAL_PROPERTIES_FLAG,
   SchemaUtilsType,
   shouldRender,
   SUBMIT_BTN_OPTIONS_KEY,
@@ -766,65 +762,6 @@ export default class Form<
     return { ...mergedErrors, schemaValidationErrors, schemaValidationErrorSchema };
   }
 
-  /** Returns the `formData` with only the elements specified in the `fields` list
-   *
-   * @param formData - The data for the `Form`
-   * @param fields - The fields to keep while filtering
-   */
-  getUsedFormData = (formData: T | undefined, fields: string[]): T | undefined => {
-    // For the case of a single input form
-    if (fields.length === 0 && typeof formData !== 'object') {
-      return formData;
-    }
-
-    const data: GenericObjectType = _pick(formData, fields);
-    if (Array.isArray(formData)) {
-      return Object.keys(data).map((key: string) => data[key]) as unknown as T;
-    }
-
-    return data as T;
-  };
-
-  /** Returns the list of field names from inspecting the `pathSchema` as well as using the `formData`
-   *
-   * @param pathSchema - The `PathSchema` object for the form
-   * @param [formData] - The form data to use while checking for empty objects/arrays
-   */
-  getFieldNames = (pathSchema: PathSchema<T>, formData?: T): string[][] => {
-    const formValueHasData = (value: T, isLeaf: boolean) =>
-      typeof value !== 'object' || _isEmpty(value) || (isLeaf && !_isEmpty(value));
-    const getAllPaths = (_obj: GenericObjectType, acc: string[][] = [], paths: string[][] = [[]]) => {
-      const objKeys = Object.keys(_obj);
-      objKeys.forEach((key: string) => {
-        const data = _obj[key];
-        if (typeof data === 'object') {
-          const newPaths = paths.map((path) => [...path, key]);
-          // If an object is marked with additionalProperties, all its keys are valid
-          if (data[RJSF_ADDITIONAL_PROPERTIES_FLAG] && data[NAME_KEY] !== '') {
-            acc.push(data[NAME_KEY]);
-          } else {
-            getAllPaths(data, acc, newPaths);
-          }
-        } else if (key === NAME_KEY && data !== '') {
-          paths.forEach((path) => {
-            const formValue = _get(formData, path);
-            const isLeaf = objKeys.length === 1;
-            // adds path to fieldNames if it points to a value or an empty object/array which is not a leaf
-            if (
-              formValueHasData(formValue, isLeaf) ||
-              (Array.isArray(formValue) && formValue.every((val) => formValueHasData(val, isLeaf)))
-            ) {
-              acc.push(path);
-            }
-          });
-        }
-      });
-      return acc;
-    };
-
-    return getAllPaths(pathSchema);
-  };
-
   /** Returns the `formData` after filtering to remove any extra data not in a form field
    *
    * @param formData - The data for the `Form`
@@ -832,13 +769,7 @@ export default class Form<
    */
   omitExtraData = (formData?: T): T | undefined => {
     const { schema, schemaUtils } = this.state;
-    const retrievedSchema = schemaUtils.retrieveSchema(schema, formData);
-    const pathSchema = schemaUtils.toPathSchema(retrievedSchema, '', formData);
-    const fieldNames = this.getFieldNames(pathSchema, formData);
-    const lodashFieldNames = fieldNames.map((fieldPaths: string[]) =>
-      Array.isArray(fieldPaths) ? fieldPaths.join('.') : fieldPaths,
-    );
-    return this.getUsedFormData(formData, lodashFieldNames);
+    return schemaUtils.omitExtraData(schema, formData);
   };
 
   /** Allows a user to set a value for the provided `fieldPath`, which must be either a dotted path to the field OR a